backend/requirements.txt
flask
flask-bcrypt
flask-cors
flask-session
python-dotenv
redis
wheel 
mysql-connector-python
flask-mysqldb 
discord